# Notify caller when apply has finished