Signed-off-by: zhangjinpeng1987 <zhangjinpeng@pingcap.com>

<!--
Thank you for contributing to TiKV! Please read TiKV's [CONTRIBUTING](https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

## What have you changed? (mandatory)

Previous behavior for prewrite/commit request:
1) scheduler receive a request and acquire latches.
2) scheduler send request to local reader ask for a snapshot.
3) local reader fetch a snapshot and redirect request to scheduler-worker-pool.
4) scheduler-worker-pool check txn constraints and generate modifications and send to raftstore.
5) after raftstore(apply worker) has applied the modifications, raftstore send a result back to scheduler.
6) scheduler release latches and notify the caller.

This branch change step 5, raftstore notify the caller and send a message to scheduler after modifications has been applied. So the latency for prewrite/commit can be reduced.

## What are the type of the changes? (mandatory)

The currently defined types are listed below, please pick one of the types for this PR by removing the others:
- Improvement (change which is an improvement to an existing feature)

## How has this PR been tested? (mandatory)

Unit tests

## Does this PR affect documentation (docs) update? (mandatory)

No

## Does this PR affect tidb-ansible update? (mandatory)

No

- [x] Benchmark

insert threads | master (QPS/avg latency/.99 latency) | branch
-- | -- | --
1 | 526 / 1.89 / 2.43 | 541 / 1.84 / 2.35
2 | 1211 / 1.65 / 2.18 | 1239 / 1.61 / 2.52
4 | 2655 / 1.5 / 2.22 | 2651 / 1.51 / 2.11
16 | 7792 / 2.05 / 4.33 | 7844 / 2.04 / 4.33
32 | 10482 / 3.05 / 8.74 | 10499 / 3.04 / 8.38

![apply-notify-cpu](https://user-images.githubusercontent.com/19700528/54864929-f3133200-4d98-11e9-8e03-0d1038a2d23f.jpeg)
As showed above:

- When the insert load is low, this feature can reduce latency by about 2%.
- Because the callbacks are involved by the apply pool, so the scheduler CPU reduced by 10%, and this part of cost transferred to apply pool, i think this is helpful when the single thread scheduler is the bottleneck.