@siddontang 2019-03-26
> i think this is helpful when the single thread scheduler is the bottleneck.

I think we will remove the single thread scheduler later, maybe you can let @hicqu benchmark in his PR.